Se habilitó el botón para salir

+++ ProyectoFontaine/Assets/GameManager/Calificacion.cs
@@ -7,10 +7,10 @@ public class Calificacion
     public int idCalificacion;
     public int idPregunta;
     public string idEstudiante;
-	public int nota;
+	public float nota;
 	public int tiempo;
 
-    public Calificacion(int idCalificacion, int idPregunta, string idEstudiante, int nota, int tiempo)
+    public Calificacion(int idCalificacion, int idPregunta, string idEstudiante, float nota, int tiempo)
     {
 		this.idCalificacion = idCalificacion;
 		this.idPregunta = idPregunta;
